App now defaults to Spanish, which is also now listed as the first language option, followed by Zapotec, followed by English. In working on this, we have discovered that Update does not work in Spanish, which is a problem.
This is a Janet-approved color scheme. We still need to find where to pass colors to the word of the day text, the actionbar text, the drawer text, and the semantic domain spinner background and text.

We also of course still need to fix the semantic domain spinner such that it works in every language, and ideally fix the download in DictionaryDatabase so you don't need to Update to get the full dictionary after downloading.